+++ core/src/core/command/Command.ts
@@ -72,6 +72,7 @@ export class Command {
   public executeReplace: CommandAdapt['replace'];
   public executePrint: CommandAdapt['print'];
   public executeExportDocx: CommandAdapt['exportDocx'];
+  public executeExportPdf: CommandAdapt['exportPdf'];
   public executeReplaceImageElement: CommandAdapt['replaceImageElement'];
   public executeSaveAsImageElement: CommandAdapt['saveAsImageElement'];
   public executeSetImageCrop: CommandAdapt['setImageCrop'];
@@ -225,6 +226,7 @@ export class Command {
     this.executeReplace = adapt.replace.bind(adapt);
     this.executePrint = adapt.print.bind(adapt);
     this.executeExportDocx = adapt.exportDocx.bind(adapt);
+    this.executeExportPdf = adapt.exportPdf.bind(adapt);
     this.executeReplaceImageElement = adapt.replaceImageElement.bind(adapt);
     this.executeSaveAsImageElement = adapt.saveAsImageElement.bind(adapt);
     this.executeSetImageCrop = adapt.setImageCrop.bind(adapt);