+++ react/src/EditorToolbar.tsx
@@ -30,7 +30,7 @@ import JustifyTool from './toolbar/JustifyTool';
 // import RowMarginTool from "./toolbar/RowMarginTool";
 import ListTool from './toolbar/ListTool';
 import ImageTool from './toolbar/ImageTool';
-import ExportDocxTool from './toolbar/ExportDocxTool';
+import ExportTool from './toolbar/ExportTool';
 // import HyperlinkTool from "./toolbar/HyperlinkTool";
 import SeparatorTool from './toolbar/SeparatorTool';
 import InsertElementTool from './toolbar/WatermarkTool';
@@ -151,7 +151,7 @@ export default function EditorToolbar() {
       {/* Media */}
       <div className="menu-item">
         <ImageTool />
-        <ExportDocxTool />
+        <ExportTool />
       </div>
       <div className="menu-item">
         <InsertElementTool />